requirements.txt
altgraph==0.17.4
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3
kiwisolver==1.4.8
matplotlib==3.10.0
networkx==3.4.2
numpy==2.2.2
packaging==24.2
pefile==2023.2.7
pillow==11.1.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pywin32==308
pywin32-ctypes==0.2.3
setuptools==75.8.0
six==1.17.0
wcwidth==0.2.13
webencodings==0.5.1
